good to remove setInitialState yay!
