+++ packages/studio-ui/src/index.ts
@@ -1,4 +1,2 @@
 export { default as App } from "./App"
 export { default as hotReloadStore } from "./store/hotReloadStore"
-export { default as generateTailwindSafelist } from "@yext/studio-plugin/src/utils/generateTailwindSafelist";
-export { default as setInitialState } from "./store/setInitialState";
