+++ packages/studio-ui/src/store/slices/createPreviousSaveSlice.ts
@@ -1,13 +1,15 @@
-
+import initialStudioData from "virtual_yext-studio";
 import PreviousSaveSlice, {
   PreviousSaveSliceState,
 } from "../models/slices/PreviousSaveSlice";
 import { SliceCreator } from "../models/utils";
 
 const createPreviousSaveSlice: SliceCreator<PreviousSaveSlice> = (set) => ({
-  siteSettings: {},
+  siteSettings: {
+    values: initialStudioData.siteSettings?.values,
+  },
   fileMetadatas: {
-    UUIDToFileMetadata: {},
+    UUIDToFileMetadata: initialStudioData.UUIDToFileMetadata,
   },
   setPreviousSave(saveState: PreviousSaveSliceState) {
     set(saveState);
