+++ packages/studio-ui/src/store/slices/createSiteSettingsSlice.ts
@@ -1,10 +1,10 @@
-
+import initialStudioData from "virtual_yext-studio";
 import SiteSettingsSlice from "../models/slices/SiteSettingsSlice";
 import { SliceCreator } from "../models/utils";
 
 const createSiteSettingsSlice: SliceCreator<SiteSettingsSlice> = (set) => ({
-  shape: undefined,
-  values: undefined,
+  shape: initialStudioData.siteSettings?.shape,
+  values: initialStudioData.siteSettings?.values,
   setShape: (shape) => set({ shape }),
   setValues: (values) => set({ values }),
 });
