+++ packages/studio-ui/src/store/slices/createStudioConfigSlice.ts
@@ -1,15 +1,10 @@
 import StudioConfigSlice from "../models/slices/StudioConfigSlice";
 import { SliceCreator } from "../models/utils";
+import initialStudioData from "virtual_yext-studio";
 
 const createStudioConfigSlice: SliceCreator<StudioConfigSlice> = () => ({
-  paths: {
-    components: '',
-    pages: '',
-    modules: '',
-    siteSettings: '',
-    localData: ''
-  },
-  isPagesJSRepo: false
+  paths: initialStudioData.studioConfig?.paths,
+  isPagesJSRepo: initialStudioData.studioConfig?.isPagesJSRepo,
 });
 
 export default createStudioConfigSlice;
