+++ packages/studio-ui/src/store/slices/createStudioEnvDataSlice.ts
@@ -1,8 +1,11 @@
 import StudioEnvDataSlice from "../models/slices/StudioEnvDataSlice";
 import { SliceCreator } from "../models/utils";
+import initialStudioData from "virtual_yext-studio";
+
+console.log({ initialStudioData })
 
 const createStudioEnvDataSlice: SliceCreator<StudioEnvDataSlice> = () => ({
-  isWithinCBD: false
+  isWithinCBD: initialStudioData.isWithinCBD,
 });
 
 export default createStudioEnvDataSlice;
